+++ src/components/section/section.scss
@@ -1,0 +1,64 @@
+.section {
+  text-align: center;
+  font-family: 'Poppins';
+  background: #E2F2FE;
+
+
+  &__title {
+    color: #140B34;
+    padding-top: 40px;
+    font-size: 32px;
+    font-weight: 600;
+    line-height: 48px;
+    text-align: center;
+  }
+
+  &__text {
+    color: #232323;
+    margin-left: 120px;
+    margin-right: 120px;
+    margin-top: 48px;
+    font-weight: 400;
+    font-size: 28px;
+    line-height: 36px;
+  }
+
+  &__text--b {
+    color: #2738F5;
+    font-weight: 600;
+  }
+
+  .itembox {
+    justify-content: space-evenly;
+
+    .item {
+      margin-top: 120px;
+      width: 286px;
+
+      &__title {
+        margin-top: 72px;
+        color: #140B34;
+        font-size: 24px;
+        font-weight: 600;
+        line-height: 36px;
+        text-align: center;
+      }
+
+      &__text {
+        height: 240px;
+        color: #232323;
+        margin-top: 32px;
+        font-weight: 400;
+        font-size: 20px;
+        line-height: 30px;
+        text-align: center;
+        margin-bottom: 40px;
+      }
+
+      &__text--b {
+        color: #140B34;
+        font-weight: 600;
+      }
+    }
+  }
+}
